+++ tests/integration/cases/call_variable_form_new/D_call@v2.d
@@ -1,0 +1,5 @@
+import std.json;
+void main() {
+int make_widget(T...)(T args) { return 0; }
+auto my_data = make_widget(42);
+}